+++ examples/generate_figures.py
@@ -8,6 +8,7 @@ from examples.battery_state_of_charge import run_battery_state_of_charge
 from examples.bearings_only import run_bearings_only_tracking
 from examples.heading_coupled import run_heading_coupled_tracking
 from examples.linear_tracking import run_planar_tracking
+from examples.known_correspondence_slam import run_known_correspondence_slam
 from examples.plotting import (
     plot_ballistic_tracking,
     plot_bearings_only_tracking,
@@ -24,6 +25,7 @@ from examples.rigid_body import (
     run_inertia_estimation_rigid_body,
     run_known_inertia_rigid_body,
 )
+from examples.slam_plotting import plot_known_correspondence_slam
 from examples.tdoa_emitter_localization import run_tdoa_emitter_localization
 from examples.tdoa_plotting import plot_tdoa_emitter_localization
 
@@ -68,6 +70,10 @@ def main():
     plot_battery_state_of_charge(
         run_battery_state_of_charge(),
         figure_directory / "battery-state-of-charge.png",
+    )
+    plot_known_correspondence_slam(
+        run_known_correspondence_slam(),
+        figure_directory / "known-correspondence-slam.png",
     )
 
     plot_rigid_body_tracking(